+++ src/main/java/blackjack/controller/BlackjackController.java
@@ -5,9 +5,9 @@ import blackjack.domain.card.exception.NoMoreCardException;
 import blackjack.domain.exception.CustomException;
 import blackjack.domain.game.BlackjackGame;
 import blackjack.domain.game.BlackjackResult;
-import blackjack.domain.player.Dealer;
-import blackjack.domain.player.Participants;
-import blackjack.domain.player.Player;
+import blackjack.domain.user.Dealer;
+import blackjack.domain.user.Participants;
+import blackjack.domain.user.Player;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 import java.util.List;
