+++ src/main/java/blackjack/domain/user/Dealer.java
@@ -1,4 +1,4 @@
-package blackjack.domain.player;
+package blackjack.domain.user;
 
 public class Dealer extends Participant {
