+++ src/main/java/blackjack/domain/user/Participant.java
@@ -1,4 +1,4 @@
-package blackjack.domain.player;
+package blackjack.domain.user;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.PlayerCards;
