+++ src/main/java/blackjack/view/OutputView.java
@@ -5,10 +5,10 @@ import static java.text.MessageFormat.format;
 import blackjack.constants.ErrorCode;
 import blackjack.domain.game.BlackjackResult;
 import blackjack.domain.game.GameResult;
-import blackjack.domain.player.Dealer;
-import blackjack.domain.player.Participant;
-import blackjack.domain.player.Participants;
-import blackjack.domain.player.Player;
+import blackjack.domain.user.Dealer;
+import blackjack.domain.user.Participant;
+import blackjack.domain.user.Participants;
+import blackjack.domain.user.Player;
 import blackjack.view.message.CardMessageConvertor;
 import blackjack.view.message.ErrorCodeMessage;
 import blackjack.view.message.GameResultMessage;
