+++ src/test/java/blackjack/domain/game/BlackjackGameTest.java
@@ -6,8 +6,8 @@ import blackjack.domain.card.Card;
 import blackjack.domain.card.CardDeck;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardSuit;
-import blackjack.domain.player.Participants;
-import blackjack.domain.player.Player;
+import blackjack.domain.user.Participants;
+import blackjack.domain.user.Player;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
