+++ src/test/java/blackjack/domain/user/NameTest.java
@@ -1,9 +1,9 @@
-package blackjack.domain.player;
+package blackjack.domain.user;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import blackjack.domain.player.exception.InvalidPlayerNameException;
+import blackjack.domain.user.exception.InvalidPlayerNameException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
