+++ src/test/java/blackjack/domain/user/ParticipantsTest.java
@@ -1,10 +1,10 @@
-package blackjack.domain.player;
+package blackjack.domain.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import blackjack.domain.player.exception.DuplicatePlayerNameException;
+import blackjack.domain.user.exception.DuplicatePlayerNameException;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
